update tests cases into own file
