+++ src/__tests__/useForm.test.tsx
@@ -9,18 +9,10 @@ import {
 } from '@testing-library/react';
 import { perf, wait, PerfTools } from 'react-performance-testing';
 import 'jest-performance-testing';
-import get from '../utils/get';
+import isFunction from '../utils/isFunction';
 import { useForm } from '../';
 import { VALIDATION_MODE } from '../constants';
-import {
-  NestedValue,
-  UseFormMethods,
-  ErrorOption,
-  FieldError,
-  RegisterOptions,
-  DeepMap,
-} from '../types';
-import isFunction from '../utils/isFunction';
+import { NestedValue, UseFormMethods, RegisterOptions } from '../types';
 
 let nodeEnv: string | undefined;
 
@@ -34,294 +26,6 @@ describe('useForm', () => {
     jest.resetAllMocks();
     jest.restoreAllMocks();
     process.env.NODE_ENV = nodeEnv;
-  });
-
-  describe('register', () => {
-    it('should return throw error when name is missing', () => {
-      process.env.NODE_ENV = 'development';
-      const { result } = renderHook(() => useForm<{ test: string }>());
-
-      // @ts-expect-error
-      const registerCallback = () => result.current.register();
-
-      expect(registerCallback).toThrow(
-        new Error(
-          '📋 `name` prop is missing during register: https://react-hook-form.com/api#register',
-        ),
-      );
-    });
-
-    it('should support register passed to ref', async () => {
-      const { result } = renderHook(() =>
-        useForm<{ test: string }>({
-          defaultValues: {
-            test: 'testData',
-          },
-        }),
-      );
-
-      const { ref } = result.current.register('test');
-
-      isFunction(ref) &&
-        ref({
-          target: {
-            value: 'testData',
-          },
-        });
-
-      await act(async () => {
-        await result.current.handleSubmit((data) => {
-          expect(data).toEqual({
-            test: 'testData',
-          });
-        })({
-          preventDefault: () => {},
-          persist: () => {},
-        } as React.SyntheticEvent);
-      });
-    });
-
-    // test.each([['text'], ['radio'], ['checkbox']])(
-    //   'should register field for %s type',
-    //   async (type) => {
-    //     const mockListener = jest.spyOn(
-    //       findRemovedFieldAndRemoveListener,
-    //       'default',
-    //     );
-    //     jest.spyOn(HTMLInputElement.prototype, 'addEventListener');
-    //
-    //     const Component = () => {
-    //       const {
-    //         register,
-    //         formState: { isDirty },
-    //       } = useForm<{
-    //         test: string;
-    //       }>();
-    //       return (
-    //         <div>
-    //           <input type={type} {...register('test')} />
-    //           <span role="alert">{`${isDirty}`}</span>
-    //         </div>
-    //       );
-    //     };
-    //
-    //     const { renderCount } = perf<{ Component: unknown }>(React);
-    //
-    //     render(<Component />);
-    //
-    //     const ref = screen.getByRole(type === 'text' ? 'textbox' : type);
-    //
-    //     expect(ref.addEventListener).toHaveBeenCalledWith(
-    //       type === 'radio' || type === 'checkbox'
-    //         ? EVENTS.CHANGE
-    //         : EVENTS.INPUT,
-    //       expect.any(Function),
-    //     );
-    //
-    //     // check MutationObserver
-    //     ref.remove();
-    //
-    //     await waitFor(() => expect(mockListener).toHaveBeenCalled());
-    //     expect(screen.getByRole('alert').textContent).toBe('false');
-    //     await wait(() =>
-    //       expect(renderCount.current.Component).toBeRenderedTimes(2),
-    //     );
-    //   },
-    // );
-
-    test.each([['text'], ['radio'], ['checkbox']])(
-      'should not register the same %s input',
-      async (type) => {
-        const callback = jest.fn();
-        const Component = () => {
-          const { register, handleSubmit } = useForm<{
-            test: string;
-          }>();
-          return (
-            <div>
-              <input type={type} {...register('test')} />
-
-              <button onClick={handleSubmit(callback)}>submit</button>
-            </div>
-          );
-        };
-
-        render(<Component />);
-
-        fireEvent.click(screen.getByRole('button', { name: /submit/ }));
-
-        await waitFor(() =>
-          expect(callback).toHaveBeenCalledWith(
-            {
-              test: type === 'checkbox' ? false : '',
-            },
-            expect.any(Object),
-          ),
-        );
-      },
-    );
-
-    it('should re-render if errors occurred with resolver when formState.isValid is defined', async () => {
-      const resolver = async (data: any) => {
-        return {
-          values: data,
-          errors: {
-            test: {
-              type: 'test',
-            },
-          },
-        };
-      };
-
-      const Component = () => {
-        const { register, formState } = useForm<{ test: string }>({
-          resolver,
-        });
-
-        return (
-          <div>
-            <input {...register('test')} />
-            <span role="alert">{`${formState.isValid}`}</span>
-          </div>
-        );
-      };
-
-      const { renderCount } = perf<{ Component: unknown }>(React);
-
-      render(<Component />);
-
-      await wait(() => expect(renderCount.current.Component).toBeMounted());
-      expect(screen.getByRole('alert').textContent).toBe('false');
-    });
-
-    it('should be set default value when item is remounted again', async () => {
-      const { result, unmount } = renderHook(() => useForm<{ test: string }>());
-
-      result.current.register('test');
-
-      result.current.setValue('test', 'test');
-
-      unmount();
-
-      const ref = { type: 'text', name: 'test' };
-
-      result.current.register('test');
-
-      expect(ref).toEqual({ type: 'text', name: 'test' });
-
-      expect(result.current.getValues()).toEqual({ test: 'test' });
-    });
-
-    // issue: https://github.com/react-hook-form/react-hook-form/issues/2298
-    it('should reset isValid formState after reset with valid value in initial render', async () => {
-      const Component = () => {
-        const { register, reset, formState } = useForm<{
-          issue: string;
-          test: string;
-        }>({
-          mode: VALIDATION_MODE.onChange,
-        });
-
-        React.useEffect(() => {
-          setTimeout(() => {
-            reset({ issue: 'test', test: 'test' });
-          });
-        }, [reset]);
-
-        return (
-          <div>
-            <input {...register('test', { required: true })} />
-            <input type="text" {...register('issue', { required: true })} />
-            <button disabled={!formState.isValid}>submit</button>
-          </div>
-        );
-      };
-
-      await actComponent(async () => {
-        render(<Component />);
-      });
-
-      await waitFor(() => {
-        expect(screen.getByRole('button')).not.toBeDisabled();
-      });
-    });
-  });
-
-  describe('register valueAs', () => {
-    it('should return number value with valueAsNumber', async () => {
-      let output = {};
-      const Component = () => {
-        const { register, handleSubmit } = useForm<{
-          test: number;
-          test1: boolean;
-        }>();
-
-        return (
-          <form onSubmit={handleSubmit((data) => (output = data))}>
-            <input {...register('test', { valueAsNumber: true })} />
-            <input
-              {...register('test1', {
-                setValueAs: (value: string) => value === 'true',
-              })}
-            />
-            <button>submit</button>
-          </form>
-        );
-      };
-
-      render(<Component />);
-
-      fireEvent.input(screen.getAllByRole('textbox')[0], {
-        target: {
-          value: '12345',
-        },
-      });
-
-      fireEvent.input(screen.getAllByRole('textbox')[1], {
-        target: {
-          value: 'true',
-        },
-      });
-
-      await actComponent(async () => {
-        await fireEvent.click(screen.getByRole('button'));
-      });
-
-      expect(output).toEqual({ test: 12345, test1: true });
-    });
-  });
-
-  describe('unregister', () => {
-    it('should unregister an registered item', async () => {
-      const { result } = renderHook(() => useForm<{ input: string }>());
-
-      result.current.register('input');
-
-      await act(async () => {
-        await result.current.unregister('input');
-      });
-
-      expect(result.current.getValues()).toEqual({});
-    });
-
-    it('should unregister an registered item with array name', async () => {
-      const { result } = renderHook(() =>
-        useForm<{
-          input: string;
-          input2: string;
-        }>(),
-      );
-
-      result.current.register('input');
-      result.current.register('input');
-      result.current.register('input2');
-
-      await act(async () => {
-        await result.current.unregister(['input', 'input2']);
-      });
-
-      expect(result.current.getValues()).toEqual({});
-    });
   });
 
   describe('when component unMount', () => {
@@ -458,2414 +162,6 @@ describe('useForm', () => {
 
       expect(formState.dirty.test).toBeDefined();
       expect(formState.isDirty).toBeTruthy();
-    });
-  });
-
-  describe('watch', () => {
-    it('should return undefined when input gets unregister', async () => {
-      const Component = () => {
-        const { register, watch, unregister } = useForm<{ test: string }>();
-        const data = watch('test');
-
-        return (
-          <>
-            <input {...register('test')} />
-            <span>{data}</span>
-            <button type="button" onClick={() => unregister('test')}>
-              hide
-            </button>
-          </>
-        );
-      };
-
-      render(<Component />);
-
-      fireEvent.input(screen.getByRole('textbox'), {
-        target: {
-          value: 'test',
-        },
-      });
-
-      screen.getByText('test');
-
-      await actComponent(async () => {
-        await fireEvent.click(screen.getByRole('button'));
-      });
-
-      expect(screen.queryByText('test')).toBeNull();
-    });
-
-    it('should watch individual input', async () => {
-      const { result } = renderHook(() => {
-        return useForm<{ test: string }>({
-          defaultValues: {
-            test: 'data',
-          },
-        });
-      });
-
-      expect(result.current.watch('test')).toBe(undefined);
-
-      result.current.register('test');
-
-      await act(async () => {
-        result.current.setValue('test', 'data1');
-      });
-
-      act(() => {
-        expect(result.current.watch('test')).toBe('data1');
-      });
-    });
-
-    it('should return default value if field is undefined', () => {
-      renderHook(() => {
-        const { watch } = useForm<{ test: string }>({
-          defaultValues: { test: 'test' },
-        });
-
-        expect(watch()).toEqual({ test: 'test' });
-      });
-    });
-
-    it('should return default value if value is empty', () => {
-      renderHook(() => {
-        const { watch } = useForm<{ test: string }>();
-
-        expect(watch('test', 'default')).toBe('default');
-      });
-    });
-
-    it('should watch array of inputs', () => {
-      const { result } = renderHook(() =>
-        useForm<{ test: string; test1: string }>(),
-      );
-
-      expect(result.current.watch(['test', 'test1'])).toEqual([
-        undefined,
-        undefined,
-      ]);
-
-      const { ref } = result.current.register('test');
-      isFunction(ref) &&
-        ref({
-          name: 'test',
-          value: 'data1',
-        });
-
-      const { ref: ref1 } = result.current.register('test1');
-      isFunction(ref1) &&
-        ref1({
-          name: 'test1',
-          value: 'data2',
-        });
-
-      expect(result.current.watch(['test', 'test1'])).toEqual([
-        'data1',
-        'data2',
-      ]);
-    });
-
-    it('should watch every fields', () => {
-      const { result } = renderHook(() =>
-        useForm<{ test: string; test1: string }>(),
-      );
-
-      const { ref } = result.current.register('test');
-      isFunction(ref) &&
-        ref({
-          name: 'test',
-          value: 'data1',
-        });
-
-      const { ref: ref1 } = result.current.register('test1');
-      isFunction(ref1) &&
-        ref1({
-          name: 'test1',
-          value: 'data2',
-        });
-
-      expect(result.current.watch()).toEqual({ test: 'data1', test1: 'data2' });
-    });
-  });
-
-  describe('reset', () => {
-    it('should reset the form and re-render the form', async () => {
-      const { result } = renderHook(() => useForm<{ test: string }>());
-
-      result.current.register('test');
-      result.current.setValue('test', 'data');
-
-      expect(result.current.formState.isSubmitted).toBeFalsy();
-      await act(async () => {
-        await result.current.handleSubmit((data) => {
-          expect(data).toEqual({
-            test: 'data',
-          });
-        })({
-          preventDefault: () => {},
-          persist: () => {},
-        } as React.SyntheticEvent);
-      });
-
-      expect(result.current.formState.isSubmitted).toBeTruthy();
-      act(() => result.current.reset());
-      expect(result.current.formState.isSubmitted).toBeFalsy();
-    });
-
-    it('should reset form value', () => {
-      let methods: any;
-      const Component = () => {
-        methods = useForm<{
-          test: string;
-        }>();
-        return (
-          <form>
-            <input {...methods.register('test')} />
-          </form>
-        );
-      };
-      render(<Component />);
-
-      actComponent(() =>
-        methods.reset({
-          test: 'test',
-        }),
-      );
-
-      expect(methods.getValues()).toEqual({
-        test: 'test',
-      });
-    });
-
-    it('should reset the form if ref is HTMLElement and parent element is form', async () => {
-      const mockReset = jest.spyOn(window.HTMLFormElement.prototype, 'reset');
-      let methods: UseFormMethods<{ test: string }>;
-      const Component = () => {
-        methods = useForm<{ test: string }>();
-        return (
-          <form>
-            <input {...methods.register('test')} />
-          </form>
-        );
-      };
-      render(<Component />);
-
-      actComponent(() => methods.reset());
-
-      expect(mockReset).toHaveBeenCalled();
-    });
-
-    it('should set array value of multiple checkbox inputs correctly', async () => {
-      const Component = () => {
-        const { register } = useForm<{
-          test: NestedValue<string[]>;
-        }>({
-          defaultValues: {
-            test: ['1', '2'],
-          },
-        });
-
-        return (
-          <>
-            <input type="checkbox" value={'1'} {...register('test')} />
-            <input type="checkbox" value={'2'} {...register('test')} />
-          </>
-        );
-      };
-
-      render(<Component />);
-
-      actComponent(() => {
-        screen
-          .getAllByRole('checkbox')
-          .forEach((checkbox) =>
-            expect((checkbox as HTMLInputElement).checked).toBeTruthy(),
-          );
-      });
-    });
-
-    it('should reset the form if ref is HTMLElement and parent element is not form', async () => {
-      const mockReset = jest.spyOn(window.HTMLFormElement.prototype, 'reset');
-      let methods: UseFormMethods<{
-        test: string;
-      }>;
-      const Component = () => {
-        methods = useForm<{
-          test: string;
-        }>();
-        return <input {...methods.register('test')} />;
-      };
-      render(<Component />);
-
-      actComponent(() => methods.reset());
-
-      expect(mockReset).not.toHaveBeenCalled();
-    });
-
-    it('should set default value if values is specified to first argument', async () => {
-      const { result } = renderHook(() =>
-        useForm<{
-          test: string;
-        }>(),
-      );
-
-      result.current.register('test');
-
-      act(() => result.current.reset({ test: 'test' }));
-
-      expect(result.current.control.defaultValuesRef.current).toEqual({
-        test: 'test',
-      });
-    });
-
-    it('should reset unmountFieldsState value when shouldUnregister set to false', () => {
-      const { result } = renderHook(() =>
-        useForm<{
-          test: string;
-        }>(),
-      );
-
-      result.current.register('test');
-
-      act(() => result.current.reset({ test: 'test' }));
-    });
-
-    it('should not reset unmountFieldsState value by default', () => {
-      const { result } = renderHook(() =>
-        useForm<{
-          test: string;
-        }>(),
-      );
-
-      result.current.register('test');
-
-      act(() => result.current.reset({ test: 'test' }));
-    });
-
-    it('should not reset form values when keepValues is specified', () => {
-      const Component = () => {
-        const { register, reset } = useForm();
-
-        return (
-          <>
-            <input {...register('test')} />
-            <button
-              type={'button'}
-              onClick={() =>
-                reset(undefined, {
-                  keepValues: true,
-                })
-              }
-            >
-              reset
-            </button>
-          </>
-        );
-      };
-
-      render(<Component />);
-
-      fireEvent.change(screen.getByRole('textbox'), {
-        target: {
-          value: 'test',
-        },
-      });
-
-      fireEvent.click(screen.getByRole('button'));
-
-      expect((screen.getByRole('textbox') as HTMLInputElement).value).toEqual(
-        'test',
-      );
-    });
-
-    it('should not reset form defaultValues when keepDefaultValues is specified', () => {
-      const Component = () => {
-        const {
-          register,
-          reset,
-          formState: { isDirty },
-        } = useForm({
-          defaultValues: {
-            test: 'test1',
-          },
-        });
-
-        return (
-          <>
-            <input {...register('test')} />
-            <p>{isDirty ? 'dirty' : ''}</p>
-            <button
-              type={'button'}
-              onClick={() =>
-                reset(undefined, {
-                  keepValues: true,
-                })
-              }
-            >
-              reset
-            </button>
-          </>
-        );
-      };
-
-      render(<Component />);
-
-      fireEvent.change(screen.getByRole('textbox'), {
-        target: {
-          value: 'test',
-        },
-      });
-
-      fireEvent.click(screen.getByRole('button'));
-
-      expect((screen.getByRole('textbox') as HTMLInputElement).value).toEqual(
-        'test',
-      );
-
-      fireEvent.change(screen.getByRole('textbox'), {
-        target: {
-          value: 'test2',
-        },
-      });
-
-      act(() => {
-        screen.getByText('dirty');
-      });
-
-      fireEvent.change(screen.getByRole('textbox'), {
-        target: {
-          value: 'test1',
-        },
-      });
-
-      expect(screen.queryByText('dirty')).toBeNull();
-    });
-
-    it('should not reset if keepStateOption is specified', async () => {
-      const { result } = renderHook(() => useForm<{ test: string }>());
-
-      result.current.register('test');
-
-      result.current.formState.errors = {
-        test: {
-          type: 'test',
-          message: 'something wrong',
-        },
-      };
-      result.current.control.validFieldsRef.current = {
-        test: true,
-      };
-      result.current.control.fieldsWithValidationRef.current = {
-        test: true,
-      };
-
-      result.current.formState.touched = { test: true };
-      result.current.formState.isDirty = true;
-      result.current.formState.isSubmitted = true;
-
-      act(() =>
-        result.current.reset(
-          { test: '' },
-          {
-            keepErrors: true,
-            keepIsDirty: true,
-            keepIsSubmitted: true,
-            keepTouched: true,
-            keepIsValid: true,
-            keepSubmitCount: true,
-            keepDirty: true,
-          },
-        ),
-      );
-
-      expect(result.current.formState.errors).toEqual({
-        test: {
-          type: 'test',
-          message: 'something wrong',
-        },
-      });
-      expect(result.current.formState.touched).toEqual({
-        test: true,
-      });
-      expect(result.current.control.validFieldsRef.current).toEqual({
-        test: true,
-      });
-      expect(result.current.control.fieldsWithValidationRef.current).toEqual({
-        test: true,
-      });
-      expect(result.current.formState.isDirty).toBeTruthy();
-      expect(result.current.formState.isSubmitted).toBeTruthy();
-    });
-  });
-
-  describe('setValue', () => {
-    it('should not setValue for unmounted state with shouldUnregister', () => {
-      const { result } = renderHook(() => useForm<{ test1: string }>());
-
-      result.current.register('test1');
-      result.current.setValue('test1', 'data');
-    });
-
-    it('should empty string when value is null or undefined when registered field is HTMLElement', () => {
-      const { result } = renderHook(() =>
-        useForm<{ test: string }>({
-          defaultValues: {
-            test: 'test',
-          },
-        }),
-      );
-
-      const elm = document.createElement('input');
-      elm.type = 'text';
-      elm.name = 'test';
-
-      result.current.register('test');
-
-      result.current.setValue('test', null);
-
-      expect(elm).not.toHaveValue();
-
-      result.current.setValue('test', undefined);
-
-      expect(elm).not.toHaveValue();
-    });
-
-    it('should set value of radio input correctly', async () => {
-      const { result } = renderHook(() => useForm<{ test: string }>());
-
-      result.current.register('test');
-
-      result.current.setValue('test', '1');
-
-      await act(async () => {
-        await result.current.handleSubmit((data) => {
-          expect(data).toEqual({
-            test: '1',
-          });
-        })({
-          preventDefault: () => {},
-          persist: () => {},
-        } as React.SyntheticEvent);
-      });
-    });
-
-    it('should set value of file input correctly if value is FileList', async () => {
-      const { result } = renderHook(() => useForm<{ test: FileList }>());
-
-      result.current.register('test');
-
-      const blob = new Blob([''], { type: 'image/png' });
-      const file = blob as File;
-      const fileList = {
-        0: file,
-        1: file,
-        length: 2,
-        item: () => file,
-      };
-
-      act(() => result.current.setValue('test', fileList));
-
-      await act(async () => {
-        await result.current.handleSubmit((data) => {
-          expect(data).toEqual({
-            test: fileList,
-          });
-        })({
-          preventDefault: () => {},
-          persist: () => {},
-        } as React.SyntheticEvent);
-      });
-    });
-
-    it('should set value of multiple checkbox input correctly', async () => {
-      const { result } = renderHook(() => useForm<{ test: string }>());
-
-      const { ref } = result.current.register('test');
-
-      const elm = document.createElement('input');
-      elm.type = 'checkbox';
-      elm.name = 'test';
-      elm.value = '2';
-
-      document.body.append(elm);
-      isFunction(ref) && ref(elm);
-
-      const { ref: ref1 } = result.current.register('test');
-
-      const elm1 = document.createElement('input');
-      elm1.type = 'checkbox';
-      elm1.name = 'test';
-      elm1.value = '1';
-
-      document.body.append(elm1);
-
-      isFunction(ref1) && ref1(elm1);
-
-      result.current.setValue('test', ['1']);
-
-      await act(async () => {
-        await result.current.handleSubmit((data) => {
-          expect(data).toEqual({
-            test: ['1'],
-          });
-        })({
-          preventDefault: () => {},
-          persist: () => {},
-        } as React.SyntheticEvent);
-      });
-    });
-
-    it('should set value of single checkbox input correctly', async () => {
-      const { result } = renderHook(() => useForm<{ test: string }>());
-
-      result.current.register('test');
-
-      result.current.setValue('test', '1');
-
-      await act(async () => {
-        await result.current.handleSubmit((data) => {
-          expect(data).toEqual({
-            test: '1',
-          });
-        })({
-          preventDefault: () => {},
-          persist: () => {},
-        } as React.SyntheticEvent);
-      });
-    });
-
-    it('should set value of multiple select correctly', async () => {
-      const { result } = renderHook(() => useForm<{ test: string }>());
-      const { ref } = result.current.register('test');
-
-      isFunction(ref) &&
-        ref({
-          type: 'checkbox',
-          refs: [{}, {}],
-        });
-
-      result.current.setValue('test', ['1']);
-
-      await act(async () => {
-        await result.current.handleSubmit((data) => {
-          expect(data).toEqual({
-            test: ['1'],
-          });
-        })({
-          preventDefault: () => {},
-          persist: () => {},
-        } as React.SyntheticEvent);
-      });
-    });
-
-    it('should set object array value', () => {
-      const { result } = renderHook(() =>
-        useForm<{
-          test: {
-            one: string;
-            two: string;
-            three: string;
-          }[];
-        }>(),
-      );
-
-      result.current.register('test.0.one');
-      result.current.register('test.0.two');
-      result.current.register('test.0.three');
-
-      act(() => {
-        result.current.setValue('test', [
-          {
-            one: 'ONE',
-            two: 'TWO',
-            three: 'THREE',
-          },
-        ]);
-      });
-
-      expect(result.current.getValues()).toEqual({
-        test: [
-          {
-            one: 'ONE',
-            two: 'TWO',
-            three: 'THREE',
-          },
-        ],
-      });
-    });
-
-    it('should set unmountFieldsState value when shouldUnregister is set to false', async () => {
-      const { result } = renderHook(() =>
-        useForm<{ test: string; checkbox: string[] }>(),
-      );
-
-      act(() => {
-        result.current.setValue('test', '1');
-        result.current.setValue('checkbox', ['1', '2']);
-        result.current.setValue('test1[0]', {
-          one: 'ONE',
-          two: 'TWO',
-          three: 'THREE',
-        });
-      });
-    });
-
-    it('should set nested value correctly ', () => {
-      const { result } = renderHook(() =>
-        useForm<{
-          test1: NestedValue<string[]>;
-          test2: NestedValue<{
-            key1: string;
-            key2: number;
-          }>;
-          test3: NestedValue<
-            {
-              key1: string;
-              key2: number;
-            }[]
-          >;
-        }>(),
-      );
-
-      result.current.register('test1');
-      result.current.register('test2');
-      result.current.register('test3');
-
-      act(() => {
-        result.current.setValue('test1', ['1', '2', '3']);
-        result.current.setValue('test2', { key1: '1', key2: 2 });
-        result.current.setValue('test3', [
-          { key1: '1', key2: 2 },
-          { key1: '3', key2: 4 },
-        ]);
-      });
-
-      expect(result.current.control.fieldsRef.current['test1']).toEqual({
-        __field: {
-          ref: { name: 'test1', value: ['1', '2', '3'] },
-          name: 'test1',
-          value: ['1', '2', '3'],
-        },
-      });
-      expect(result.current.control.fieldsRef.current['test2']).toEqual({
-        __field: {
-          ref: { name: 'test2', value: { key1: '1', key2: 2 } },
-          name: 'test2',
-          value: { key1: '1', key2: 2 },
-        },
-      });
-      expect(result.current.control.fieldsRef.current['test3']).toEqual({
-        __field: {
-          ref: {
-            name: 'test3',
-            value: [
-              { key1: '1', key2: 2 },
-              { key1: '3', key2: 4 },
-            ],
-          },
-          name: 'test3',
-          value: [
-            { key1: '1', key2: 2 },
-            { key1: '3', key2: 4 },
-          ],
-        },
-      });
-    });
-
-    it('should work with array fields', () => {
-      const { result } = renderHook(() =>
-        useForm<{
-          test: string[];
-          test1: {
-            test: string;
-          }[];
-        }>(),
-      );
-
-      result.current.register('test1.0.test');
-      result.current.register('test.0');
-      result.current.register('test.1');
-      result.current.register('test.2');
-
-      act(() => result.current.setValue('test', ['1', '2', '3']));
-
-      expect(get(result.current.control.fieldsRef.current, 'test.0')).toEqual({
-        __field: {
-          ref: { name: 'test.0', value: '1' },
-          name: 'test.0',
-          value: '1',
-        },
-      });
-      expect(get(result.current.control.fieldsRef.current, 'test.1')).toEqual({
-        __field: {
-          ref: { name: 'test.1', value: '2' },
-          name: 'test.1',
-          value: '2',
-        },
-      });
-      expect(get(result.current.control.fieldsRef.current, 'test.2')).toEqual({
-        __field: {
-          ref: { name: 'test.2', value: '3' },
-          name: 'test.2',
-          value: '3',
-        },
-      });
-    });
-
-    it('should worked with nested array fields with object', () => {
-      const { result } = renderHook(() =>
-        useForm<{
-          test: {
-            test: string;
-          }[];
-        }>(),
-      );
-
-      result.current.register('test.0.test');
-      result.current.register('test.1.test');
-      result.current.register('test.2.test');
-
-      act(() =>
-        result.current.setValue('test', [
-          { test: '1' },
-          { test: '2' },
-          { test: '3' },
-        ]),
-      );
-
-      expect(
-        get(result.current.control.fieldsRef.current, 'test.0.test'),
-      ).toEqual({
-        __field: {
-          ref: { name: 'test.0.test', value: '1' },
-          name: 'test.0.test',
-          value: '1',
-        },
-      });
-      expect(
-        get(result.current.control.fieldsRef.current, 'test.1.test'),
-      ).toEqual({
-        __field: {
-          ref: { name: 'test.1.test', value: '2' },
-          name: 'test.1.test',
-          value: '2',
-        },
-      });
-      expect(
-        get(result.current.control.fieldsRef.current, 'test.2.test'),
-      ).toEqual({
-        __field: {
-          ref: { name: 'test.2.test', value: '3' },
-          name: 'test.2.test',
-          value: '3',
-        },
-      });
-    });
-
-    it('should work with object fields', () => {
-      const { result } = renderHook(() =>
-        useForm<{
-          test1: {
-            test: string;
-          }[];
-          test: {
-            bill: string;
-            luo: string;
-            test: string;
-          };
-        }>(),
-      );
-
-      result.current.register('test1.0.test');
-      result.current.register('test.bill');
-      result.current.register('test.luo');
-      result.current.register('test.test');
-
-      act(() =>
-        result.current.setValue('test', { bill: '1', luo: '2', test: '3' }),
-      );
-      expect(
-        get(result.current.control.fieldsRef.current, 'test.bill'),
-      ).toEqual({
-        __field: {
-          ref: { name: 'test.bill', value: '1' },
-          name: 'test.bill',
-          value: '1',
-        },
-      });
-      expect(get(result.current.control.fieldsRef.current, 'test.luo')).toEqual(
-        {
-          __field: {
-            ref: { name: 'test.luo', value: '2' },
-            name: 'test.luo',
-            value: '2',
-          },
-        },
-      );
-      expect(
-        get(result.current.control.fieldsRef.current, 'test.test'),
-      ).toEqual({
-        __field: {
-          ref: { name: 'test.test', value: '3' },
-          name: 'test.test',
-          value: '3',
-        },
-      });
-    });
-
-    it('should not work if field is not registered', () => {
-      const { result } = renderHook(() => useForm());
-
-      act(() => {
-        result.current.setValue('test', '1');
-      });
-
-      expect(result.current.control.fieldsRef.current['test']).toBeUndefined();
-    });
-
-    describe('with watch', () => {
-      it('should get watched value', () => {
-        const { result } = renderHook(() => {
-          const { register, watch, setValue } = useForm<{ test: string }>();
-
-          register('test');
-
-          React.useEffect(() => {
-            setValue('test', 'abc');
-          }, [setValue]);
-
-          return watch('test');
-        });
-
-        expect(result.current).toBe('abc');
-      });
-    });
-
-    describe('with validation', () => {
-      it('should be called trigger method if shouldValidate variable is true', async () => {
-        const { result } = renderHook(() =>
-          useForm<{
-            test: string;
-          }>(),
-        );
-
-        result.current.register('test', {
-          minLength: {
-            value: 5,
-            message: 'min',
-          },
-        });
-
-        result.current.formState.dirty;
-
-        await act(async () =>
-          result.current.setValue('test', 'abc', {
-            shouldValidate: true,
-          }),
-        );
-
-        expect(result.current.formState.errors?.test?.message).toBe('min');
-      });
-
-      it('should not be called trigger method if config is empty', async () => {
-        const { result } = renderHook(() => useForm<{ test: string }>());
-
-        result.current.register('test', {
-          minLength: {
-            value: 5,
-            message: 'min',
-          },
-        });
-
-        result.current.setValue('test', 'abc');
-
-        expect(result.current.formState.errors?.test).toBeUndefined();
-      });
-
-      it('should be called trigger method if shouldValidate variable is true and field value is array', async () => {
-        const { result } = renderHook(() =>
-          useForm<{
-            test: string[];
-          }>(),
-        );
-
-        const rules = {
-          minLength: {
-            value: 5,
-            message: 'min',
-          },
-        };
-
-        result.current.register('test.0', rules);
-        result.current.register('test.1', rules);
-        result.current.register('test.2', rules);
-
-        await act(async () =>
-          result.current.setValue('test', ['abc1', 'abc2', 'abc3'], {
-            shouldValidate: true,
-          }),
-        );
-
-        expect(result.current.formState.errors?.test?.[0]?.message).toBe('min');
-        expect(result.current.formState.errors?.test?.[1]?.message).toBe('min');
-        expect(result.current.formState.errors?.test?.[2]?.message).toBe('min');
-      });
-
-      it('should not be called trigger method if config is empty and field value is array', async () => {
-        const { result } = renderHook(() =>
-          useForm<{
-            test: string[];
-          }>(),
-        );
-
-        const rules = {
-          minLength: {
-            value: 5,
-            message: 'min',
-          },
-        };
-
-        result.current.register('test.0', rules);
-        result.current.register('test.1', rules);
-        result.current.register('test.2', rules);
-
-        act(() => result.current.setValue('test', ['test', 'test1', 'test2']));
-
-        expect(result.current.formState.errors?.test).toBeUndefined();
-      });
-    });
-
-    describe('with dirty', () => {
-      it.each(['isDirty', 'dirty'])(
-        'should be dirty when %s is defined when shouldDirty is true',
-        (property) => {
-          const { result } = renderHook(() => useForm<{ test: string }>());
-
-          result.current.formState[property as 'dirty' | 'isDirty'];
-          result.current.formState.isDirty;
-
-          result.current.register('test');
-
-          act(() =>
-            result.current.setValue('test', 'test', { shouldDirty: true }),
-          );
-
-          expect(result.current.formState.isDirty).toBeTruthy();
-          expect(result.current.formState.dirty).toEqual({ test: true });
-        },
-      );
-
-      it.each([
-        ['isDirty', ['test1', 'test2', 'test3'], [true, true, true]],
-        ['dirty', ['test1', 'test2', 'test3'], [true, true, true]],
-        ['isDirty', ['test1', '', 'test3'], [true, undefined, true]],
-        ['dirty', ['test1', '', 'test3'], [true, undefined, true]],
-      ])(
-        'should be dirty when %s is defined when shouldDirty is true with array fields',
-        (property, values, dirtyFields) => {
-          const { result } = renderHook(() =>
-            useForm<{
-              test: string[];
-            }>({
-              defaultValues: {
-                test: ['', '', ''],
-              },
-            }),
-          );
-
-          result.current.formState[property as 'isDirty' | 'dirty'];
-          result.current.formState.isDirty;
-
-          result.current.register('test.0');
-          result.current.register('test.1');
-          result.current.register('test.2');
-
-          act(() =>
-            result.current.setValue('test', values, {
-              shouldDirty: true,
-            }),
-          );
-
-          expect(result.current.formState.isDirty).toBeTruthy();
-          expect(result.current.formState.dirty).toEqual({
-            test: dirtyFields,
-          });
-        },
-      );
-
-      it.each(['isDirty', 'dirty'])(
-        'should not be dirty when %s is defined when shouldDirty is false',
-        (property) => {
-          const { result } = renderHook(() =>
-            useForm<{
-              test: string;
-            }>(),
-          );
-
-          result.current.formState[property as 'isDirty' | 'dirty'];
-
-          result.current.register('test');
-
-          act(() =>
-            result.current.setValue('test', 'test', { shouldDirty: false }),
-          );
-
-          expect(result.current.formState.isDirty).toBeFalsy();
-          expect(result.current.formState.dirty).toEqual({});
-        },
-      );
-
-      it.each(['isDirty', 'dirtyFields'])(
-        'should set name to dirtyFieldRef if field value is different with default value when formState.dirty is defined',
-        (property) => {
-          const { result } = renderHook(() =>
-            useForm<{ test: string }>({
-              defaultValues: { test: 'default' },
-            }),
-          );
-          result.current.formState[property as 'dirty' | 'isDirty'];
-          result.current.formState.isDirty;
-
-          result.current.register('test');
-
-          act(() =>
-            result.current.setValue('test', '1', { shouldDirty: true }),
-          );
-
-          expect(result.current.formState.isDirty).toBeTruthy();
-          expect(result.current.formState.dirty.test).toBeTruthy();
-        },
-      );
-
-      it.each(['isDirty', 'dirty'])(
-        'should unset name from dirtyFieldRef if field value is not different with default value when formState.dirty is defined',
-        (property) => {
-          const { result } = renderHook(() =>
-            useForm<{ test: string }>({
-              defaultValues: { test: 'default' },
-            }),
-          );
-          result.current.formState[property as 'isDirty' | 'dirty'];
-          result.current.formState.isDirty;
-
-          result.current.register('test');
-
-          act(() =>
-            result.current.setValue('test', '1', { shouldDirty: true }),
-          );
-
-          expect(result.current.formState.isDirty).toBeTruthy();
-          expect(result.current.formState.dirty.test).toBeTruthy();
-
-          act(() =>
-            result.current.setValue('test', 'default', { shouldDirty: true }),
-          );
-
-          expect(result.current.formState.isDirty).toBeFalsy();
-          expect(result.current.formState.dirty.test).toBeUndefined();
-        },
-      );
-    });
-
-    it('should set hidden input value correctly and reflect on the submission data', async () => {
-      let submitData = undefined;
-
-      const Component = () => {
-        const { register, handleSubmit, setValue } = useForm<{
-          test: string;
-        }>();
-
-        return (
-          <div>
-            <input type="hidden" defaultValue="test" {...register('test')} />
-            <button
-              onClick={() => {
-                setValue('test', 'changed');
-              }}
-            >
-              change
-            </button>
-            <button
-              onClick={handleSubmit((data) => {
-                submitData = data;
-              })}
-            >
-              submit
-            </button>
-          </div>
-        );
-      };
-
-      render(<Component />);
-
-      await actComponent(async () => {
-        await fireEvent.click(screen.getByRole('button', { name: 'change' }));
-      });
-
-      await actComponent(async () => {
-        await fireEvent.click(screen.getByRole('button', { name: 'submit' }));
-      });
-
-      expect(submitData).toEqual({
-        test: 'changed',
-      });
-    });
-
-    it('should validate the input and return correct isValid formState', async () => {
-      const { result } = renderHook(() =>
-        useForm<{ test: { data: string; data1: string } }>({
-          mode: VALIDATION_MODE.onChange,
-        }),
-      );
-
-      result.current.formState.isValid;
-
-      await act(async () => {
-        await result.current.register('test.data', { required: true });
-        await result.current.register('test.data1', { required: true });
-      });
-
-      await act(async () => {
-        await result.current.trigger();
-      });
-
-      await act(async () => {
-        result.current.setValue('test.data', 'test', { shouldValidate: true });
-      });
-
-      expect(result.current.formState.isValid).toBeFalsy();
-
-      await act(async () => {
-        await result.current.setValue('test.data1', 'test', {
-          shouldValidate: true,
-        });
-      });
-
-      expect(result.current.formState.isValid).toBeTruthy();
-    });
-  });
-
-  describe('trigger', () => {
-    it('should console warn when field is not found', async () => {
-      jest.spyOn(console, 'warn').mockImplementation(() => {});
-
-      process.env.NODE_ENV = 'development';
-
-      const { result } = renderHook(() => useForm<{ test: string }>());
-      expect(await result.current.trigger('test')).toBeFalsy();
-
-      expect(console.warn).toBeCalledTimes(1);
-    });
-
-    it('should remove all errors before set new errors when trigger entire form', async () => {
-      const Component = () => {
-        const [show, setShow] = React.useState(true);
-        const {
-          register,
-          unregister,
-          trigger,
-          formState: { errors },
-        } = useForm<{
-          test: string;
-        }>();
-
-        return (
-          <div>
-            {show && <input {...register('test', { required: true })} />}
-            <button type={'button'} onClick={() => trigger()}>
-              trigger
-            </button>
-            <button
-              type={'button'}
-              onClick={() => {
-                setShow(false);
-                unregister('test');
-              }}
-            >
-              toggle
-            </button>
-            {errors.test && <span>error</span>}
-          </div>
-        );
-      };
-
-      render(<Component />);
-
-      await actComponent(async () => {
-        await fireEvent.click(screen.getByRole('button', { name: 'trigger' }));
-      });
-
-      await waitFor(() => screen.getByText('error'));
-
-      await actComponent(async () => {
-        await fireEvent.click(screen.getByRole('button', { name: 'toggle' }));
-      });
-
-      await actComponent(async () => {
-        await fireEvent.click(screen.getByRole('button', { name: 'trigger' }));
-      });
-
-      expect(screen.queryByText('error')).toBeNull();
-    });
-
-    it('should return empty errors when field is found and validation pass', async () => {
-      const { result } = renderHook(() => useForm<{ test: string }>());
-      const { errors } = result.current.formState;
-
-      result.current.register('test');
-
-      await act(async () => {
-        await result.current.trigger('test');
-      });
-
-      await act(async () => {
-        await expect(errors).toEqual({});
-      });
-    });
-
-    it('should update value when value is supplied', async () => {
-      const { result } = renderHook(() => useForm<{ test: string }>());
-
-      const { errors } = result.current.formState;
-
-      result.current.register('test', { required: true });
-
-      result.current.setValue('test', 'abc');
-
-      await act(async () => {
-        await result.current.trigger('test');
-      });
-
-      await actComponent(async () => {
-        expect(errors).toEqual({});
-      });
-    });
-
-    it('should trigger multiple fields validation', async () => {
-      const { result } = renderHook(() =>
-        useForm<{ test: string; test1: string }>({
-          mode: VALIDATION_MODE.onChange,
-        }),
-      );
-
-      result.current.register('test', { required: 'required' });
-      result.current.register('test1', { required: 'required' });
-
-      await act(async () => {
-        await result.current.trigger(['test', 'test1']);
-      });
-
-      expect(result.current.formState.errors?.test?.message).toBe('required');
-      expect(result.current.formState.errors?.test1?.message).toBe('required');
-    });
-  });
-
-  describe('trigger with schema', () => {
-    it('should return the error with single field validation', async () => {
-      const resolver = async (data: any) => {
-        return {
-          values: data,
-          errors: {
-            test: {
-              type: 'test',
-            },
-          },
-        };
-      };
-
-      const { result } = renderHook(() =>
-        useForm<{ test: string }>({
-          mode: VALIDATION_MODE.onChange,
-          resolver,
-        }),
-      );
-
-      result.current.register('test', { required: true });
-
-      await act(async () => {
-        await result.current.trigger('test');
-      });
-      expect(result.current.formState.errors).toEqual({
-        test: { type: 'test' },
-      });
-    });
-
-    it('should return the status of the requested field with single field validation', async () => {
-      const resolver = async (data: any) => {
-        return {
-          values: data,
-          errors: {
-            test2: {
-              type: 'test',
-            },
-          },
-        };
-      };
-
-      const { result } = renderHook(() =>
-        useForm<{ test1: string; test2: string }>({
-          mode: VALIDATION_MODE.onChange,
-          resolver,
-        }),
-      );
-
-      result.current.register('test1', { required: false });
-      result.current.register('test2', { required: true });
-
-      await act(async () =>
-        expect(await result.current.trigger('test2')).toBeFalsy(),
-      );
-
-      expect(result.current.formState.errors).toEqual({
-        test2: {
-          type: 'test',
-        },
-      });
-    });
-
-    it('should not trigger any error when schema validation result not found', async () => {
-      const { result } = renderHook(() =>
-        useForm<{ test: string; test1: string }>({
-          mode: VALIDATION_MODE.onChange,
-          resolver: async () => {
-            return {
-              values: {},
-              errors: {
-                test: {
-                  type: 'test',
-                },
-              },
-            };
-          },
-        }),
-      );
-
-      result.current.register('test', { required: true });
-
-      await act(async () => {
-        await result.current.trigger('test1');
-      });
-
-      expect(result.current.formState.errors).toEqual({});
-    });
-
-    it('should support array of fields for schema validation', async () => {
-      const resolver = async (data: any) => {
-        return {
-          values: data,
-          errors: {
-            test1: {
-              type: 'test1',
-            },
-            test: {
-              type: 'test',
-            },
-          },
-        };
-      };
-
-      const { result } = renderHook(() =>
-        useForm<{ test: string; test1: string }>({
-          mode: VALIDATION_MODE.onChange,
-          resolver,
-        }),
-      );
-
-      result.current.register('test', { required: true });
-
-      await act(async () => {
-        await result.current.trigger(['test', 'test1']);
-      });
-
-      expect(result.current.formState.errors).toEqual({
-        test1: {
-          type: 'test1',
-        },
-        test: {
-          type: 'test',
-        },
-      });
-    });
-
-    it('should return the status of the requested fields with array of fields for validation', async () => {
-      const { result } = renderHook(() =>
-        useForm<{ test1: string; test2: string; test3: string }>({
-          mode: VALIDATION_MODE.onChange,
-          resolver: async () => {
-            return {
-              values: {},
-              errors: {
-                test3: {
-                  type: 'test',
-                },
-              },
-            };
-          },
-        }),
-      );
-
-      const { errors } = result.current.formState;
-
-      result.current.register('test1', { required: false });
-      result.current.register('test2', { required: false });
-      result.current.register('test3', { required: true });
-
-      await act(async () => {
-        await result.current.trigger(['test1', 'test2']);
-      });
-
-      await act(async () => {
-        expect(errors).toEqual({});
-      });
-
-      await act(async () => {
-        await result.current.trigger(['test3']);
-      });
-
-      await act(async () => {
-        expect(errors).toEqual({
-          test3: {
-            type: 'test',
-          },
-        });
-      });
-    });
-
-    it('should validate all fields when pass with undefined', async () => {
-      const resolver = async (data: any) => {
-        return {
-          values: data,
-          errors: {
-            test1: {
-              type: 'test1',
-            },
-            test: {
-              type: 'test',
-            },
-          },
-        };
-      };
-
-      const { result } = renderHook(() =>
-        useForm<{ test1: string; test: string }>({
-          mode: VALIDATION_MODE.onChange,
-          resolver,
-        }),
-      );
-
-      result.current.register('test', { required: true });
-      result.current.register('test1', { required: true });
-
-      await act(async () => {
-        await result.current.trigger();
-      });
-
-      expect(result.current.formState.errors).toEqual({
-        test1: {
-          type: 'test1',
-        },
-        test: {
-          type: 'test',
-        },
-      });
-    });
-  });
-
-  describe('handleSubmit', () => {
-    it('should invoke the callback when validation pass', async () => {
-      const { result } = renderHook(() => useForm());
-      const callback = jest.fn();
-
-      await act(async () => {
-        await result.current.handleSubmit(callback)({
-          preventDefault: () => {},
-          persist: () => {},
-        } as React.SyntheticEvent);
-      });
-      expect(callback).toBeCalled();
-    });
-
-    it('should pass default value', async () => {
-      const { result } = renderHook(() =>
-        useForm<{ test: string; deep: { nested: string; values: string } }>({
-          mode: VALIDATION_MODE.onSubmit,
-          defaultValues: {
-            test: 'data',
-            deep: {
-              values: '5',
-            },
-          },
-        }),
-      );
-
-      result.current.register('test');
-      result.current.register('deep.nested');
-      result.current.register('deep.values');
-
-      await act(async () => {
-        await result.current.handleSubmit((data: any) => {
-          expect(data).toEqual({
-            test: 'data',
-            deep: {
-              nested: undefined,
-              values: '5',
-            },
-          });
-        })({
-          preventDefault: () => {},
-          persist: () => {},
-        } as React.SyntheticEvent);
-      });
-    });
-
-    it('should not pass default value when field is not registered', async () => {
-      const { result } = renderHook(() =>
-        useForm<{ test: string; deep: { nested: string; values: string } }>({
-          mode: VALIDATION_MODE.onSubmit,
-          defaultValues: {
-            test: 'data',
-            deep: {
-              values: '5',
-            },
-          },
-        }),
-      );
-
-      await act(async () => {
-        await result.current.handleSubmit((data: any) => {
-          expect(data).toEqual({});
-        })({
-          preventDefault: () => {},
-          persist: () => {},
-        } as React.SyntheticEvent);
-      });
-    });
-
-    it('should invoke reRender method when readFormStateRef.current.isSubmitting is true', async () => {
-      const Component = () => {
-        const { register, handleSubmit, formState } = useForm<{
-          test: string;
-        }>();
-        return (
-          <div>
-            <input {...register('test')} />
-            <button onClick={handleSubmit(() => {})}></button>
-            <span role="alert">
-              {formState.isSubmitting ? 'true' : 'false'}
-            </span>
-          </div>
-        );
-      };
-
-      const { renderCount } = perf<{ Component: unknown }>(React);
-
-      render(<Component />);
-
-      fireEvent.click(screen.getByRole('button'));
-
-      const span = screen.getByRole('alert')!;
-      await waitFor(
-        () => {
-          if (renderCount.current.Component?.value === 2) {
-            expect(span.textContent).toBe('true');
-          } else {
-            expect(span.textContent).toBe('false');
-          }
-        },
-        { container: span },
-      );
-
-      await wait(() => {
-        expect(renderCount.current.Component).toBeRenderedTimes(4);
-      });
-    });
-
-    it('should not invoke callback when there are errors', async () => {
-      const { result } = renderHook(() => useForm<{ test: string }>());
-
-      result.current.register('test', { required: true });
-
-      const callback = jest.fn();
-
-      await act(async () => {
-        await result.current.handleSubmit(callback)({
-          preventDefault: () => {},
-          persist: () => {},
-        } as React.SyntheticEvent);
-      });
-      expect(callback).not.toBeCalled();
-    });
-
-    it('should not focus if errors is exist', async () => {
-      const focus = jest.fn();
-      const { result } = renderHook(() => useForm<{ test: string }>());
-      const { ref } = result.current.register('test', { required: true });
-
-      isFunction(ref) &&
-        ref({
-          focus,
-        });
-
-      const callback = jest.fn();
-      await act(async () => {
-        await result.current.handleSubmit(callback)({
-          preventDefault: () => {},
-          persist: () => {},
-        } as React.SyntheticEvent);
-      });
-
-      expect(callback).not.toBeCalled();
-      expect(focus).toBeCalled();
-      expect(result.current.formState.errors?.test?.type).toBe('required');
-    });
-
-    it('should not focus if shouldFocusError is false', async () => {
-      const mockFocus = jest.spyOn(HTMLInputElement.prototype, 'focus');
-
-      const { result } = renderHook(() =>
-        useForm<{ test: string }>({ shouldFocusError: false }),
-      );
-
-      result.current.register('test', { required: true });
-
-      const callback = jest.fn();
-      await act(async () => {
-        await result.current.handleSubmit(callback)({
-          preventDefault: () => {},
-          persist: () => {},
-        } as React.SyntheticEvent);
-      });
-
-      expect(callback).not.toBeCalled();
-      expect(mockFocus).not.toBeCalled();
-      expect(result.current.formState.errors?.test?.type).toBe('required');
-    });
-
-    it('should submit form data when inputs are removed', async () => {
-      const { result, unmount } = renderHook(() =>
-        useForm<{
-          test: string;
-        }>(),
-      );
-
-      result.current.register('test');
-      result.current.setValue('test', 'test');
-
-      unmount();
-
-      await act(async () =>
-        result.current.handleSubmit((data) => {
-          expect(data).toEqual({
-            test: 'test',
-          });
-        })({
-          preventDefault: () => {},
-          persist: () => {},
-        } as React.SyntheticEvent),
-      );
-    });
-
-    it('should invoke onSubmit callback and reset nested errors when submit with valid form values', async () => {
-      const callback = jest.fn();
-      const { result } = renderHook(() =>
-        useForm<{
-          test: { firstName: string; lastName: string }[];
-        }>(),
-      );
-      const validate = () => {
-        return !!result.current
-          .getValues()
-          .test.some(({ firstName }) => firstName);
-      };
-
-      result.current.register('test.0.firstName', {
-        validate,
-      });
-      result.current.register('test.0.lastName', {
-        validate,
-      });
-      result.current.register('test.1.firstName', {
-        validate,
-      });
-      result.current.register('test.1.lastName', {
-        validate,
-      });
-
-      await act(async () => {
-        await result.current.handleSubmit(callback)({
-          preventDefault: () => {},
-          persist: () => {},
-        } as React.SyntheticEvent);
-      });
-
-      expect(callback).not.toBeCalled();
-
-      result.current.setValue('test.0.firstName', 'test');
-
-      await act(async () => {
-        await result.current.handleSubmit(callback)({
-          preventDefault: () => {},
-          persist: () => {},
-        } as React.SyntheticEvent);
-      });
-
-      expect(callback).toBeCalled();
-    });
-  });
-
-  describe('handleSubmit with validationSchema', () => {
-    it('should invoke callback when error not found', async () => {
-      const resolver = async (data: any) => {
-        return {
-          values: data,
-          errors: {},
-        };
-      };
-
-      const { result } = renderHook(() =>
-        useForm<{ test: string }>({
-          mode: VALIDATION_MODE.onSubmit,
-          resolver,
-        }),
-      );
-
-      result.current.register('test', { required: true });
-
-      const callback = jest.fn();
-
-      await act(async () => {
-        await result.current.handleSubmit(callback)({
-          preventDefault: () => {},
-          persist: () => {},
-        } as React.SyntheticEvent);
-      });
-      expect(callback).toBeCalled();
-    });
-
-    it('should invoke callback with transformed values', async () => {
-      const resolver = async () => {
-        return {
-          values: { test: 'test' },
-          errors: {},
-        };
-      };
-
-      const { result } = renderHook(() =>
-        useForm<{ test: string }>({
-          mode: VALIDATION_MODE.onSubmit,
-          resolver,
-        }),
-      );
-
-      result.current.register('test', { required: true });
-
-      const callback = jest.fn();
-
-      await act(async () => {
-        await result.current.handleSubmit(callback)({
-          preventDefault: () => {},
-          persist: () => {},
-        } as React.SyntheticEvent);
-      });
-      expect(callback.mock.calls[0][0]).toEqual({ test: 'test' });
-    });
-  });
-
-  describe('handleSubmit with onInvalid callback', () => {
-    it('should invoke the onValid callback when validation pass', async () => {
-      const { result } = renderHook(() => useForm());
-      const onValidCallback = jest.fn();
-      const onInvalidCallback = jest.fn();
-
-      await act(async () => {
-        await result.current.handleSubmit(
-          onValidCallback,
-          onInvalidCallback,
-        )({
-          preventDefault: () => {},
-          persist: () => {},
-        } as React.SyntheticEvent);
-      });
-      expect(onValidCallback).toBeCalledTimes(1);
-      expect(onInvalidCallback).not.toBeCalledTimes(1);
-    });
-
-    it('should invoke the onInvalid callback when validation failed', async () => {
-      const { result } = renderHook(() =>
-        useForm<{
-          test: string;
-        }>(),
-      );
-      result.current.register('test', { required: true });
-      const onValidCallback = jest.fn();
-      const onInvalidCallback = jest.fn();
-
-      await act(async () => {
-        await result.current.handleSubmit(
-          onValidCallback,
-          onInvalidCallback,
-        )({
-          preventDefault: () => {},
-          persist: () => {},
-        } as React.SyntheticEvent);
-      });
-
-      expect(onValidCallback).not.toBeCalledTimes(1);
-      expect(onInvalidCallback).toBeCalledTimes(1);
-    });
-  });
-
-  describe('getValues', () => {
-    it('should return defaultValues before inputs mounted', () => {
-      let values;
-
-      const Component = () => {
-        const { getValues } = useForm({
-          defaultValues: {
-            test: 'test',
-          },
-        });
-
-        values = getValues();
-
-        return null;
-      };
-
-      const { rerender } = render(<Component />);
-
-      expect(values).toEqual({
-        test: 'test',
-      });
-
-      rerender(<Component />);
-
-      expect(values).toEqual({});
-    });
-
-    it('should call getFieldsValues and return all values', () => {
-      const { result } = renderHook(() => useForm<{ test: string }>());
-      result.current.register('test');
-      result.current.setValue('test', 'test');
-      expect(result.current.getValues()).toEqual({ test: 'test' });
-    });
-
-    it('should get individual field value', () => {
-      const { result } = renderHook(() =>
-        useForm<{ test: string }>({
-          defaultValues: {
-            test: '123',
-          },
-        }),
-      );
-      result.current.register('test');
-      expect(result.current.getValues('test')).toEqual('123');
-    });
-
-    it('should get all field values', () => {
-      const values = {
-        test: 'test',
-        test1: 'test1',
-        test2: 'test2',
-      };
-      const { result } = renderHook(() =>
-        useForm<{ test: string; test1: string; test2: string }>({
-          defaultValues: values,
-        }),
-      );
-      result.current.register('test');
-      result.current.register('test1');
-      result.current.register('test2');
-
-      result.current.setValue('test', 'test');
-      result.current.setValue('test1', 'test1');
-      result.current.setValue('test2', 'test2');
-
-      expect(result.current.getValues(['test', 'test1', 'test2'])).toEqual([
-        'test',
-        'test1',
-        'test2',
-      ]);
-    });
-
-    it('should get undefined when field not found', () => {
-      const { result } = renderHook(() => useForm());
-
-      expect(result.current.getValues('test')).toEqual(undefined);
-    });
-
-    it('should get value from shallowFieldsStateRef by name', () => {
-      const { result, unmount } = renderHook(() =>
-        useForm<{
-          test: string;
-        }>(),
-      );
-
-      result.current.register('test');
-      result.current.setValue('test', 'test');
-
-      unmount();
-
-      expect(result.current.getValues('test')).toEqual('test');
-    });
-
-    it('should get value from shallowFieldsStateRef by array', () => {
-      const { result, unmount } = renderHook(() =>
-        useForm<{
-          test: string;
-        }>(),
-      );
-
-      result.current.register('test');
-      result.current.setValue('test', 'test');
-
-      unmount();
-
-      expect(result.current.getValues(['test'])).toEqual(['test']);
-    });
-
-    it('should get value from shallowFieldsStateRef', () => {
-      const { result, unmount } = renderHook(() =>
-        useForm<{
-          test: string;
-        }>(),
-      );
-
-      result.current.register('test');
-      result.current.setValue('test', 'test');
-
-      unmount();
-
-      expect(result.current.getValues()).toEqual({
-        test: 'test',
-      });
-    });
-
-    it('should not get value from default value by name when field is not registered', () => {
-      const { result } = renderHook(() =>
-        useForm({
-          defaultValues: {
-            test: 'default',
-          },
-        }),
-      );
-
-      expect(result.current.getValues('test')).toBeUndefined();
-    });
-
-    it('should not get value from default value by array when field is not registered', () => {
-      const { result } = renderHook(() =>
-        useForm({
-          defaultValues: {
-            test: 'default',
-          },
-        }),
-      );
-
-      expect(result.current.getValues(['test'])).toEqual([undefined]);
-    });
-
-    it('should not get value from default value when field is not registered', () => {
-      const { result } = renderHook(() =>
-        useForm({
-          defaultValues: {
-            test: 'default',
-          },
-        }),
-      );
-
-      expect(result.current.getValues()).toEqual({});
-    });
-  });
-
-  describe('setError', () => {
-    const tests: [string, ErrorOption, DeepMap<any, FieldError>][] = [
-      [
-        'should only set an error when it is not existed',
-        { type: 'test' },
-        {
-          input: {
-            type: 'test',
-            message: undefined,
-            ref: undefined,
-          },
-        },
-      ],
-      [
-        'should set error message',
-        { type: 'test', message: 'test' },
-        {
-          input: {
-            type: 'test',
-            message: 'test',
-            ref: undefined,
-            types: undefined,
-          },
-        },
-      ],
-      [
-        'should set multiple error type',
-        {
-          types: { test1: 'test1', test2: 'test2' },
-        },
-        {
-          input: {
-            types: {
-              test1: 'test1',
-              test2: 'test2',
-            },
-            ref: undefined,
-          },
-        },
-      ],
-    ];
-    test.each(tests)('%s', (_, input, output) => {
-      const { result } = renderHook(() => useForm<{ input: string }>());
-      act(() => {
-        result.current.setError('input', input);
-      });
-      expect(result.current.formState.errors).toEqual(output);
-      expect(result.current.formState.isValid).toBeFalsy();
-    });
-  });
-
-  describe('clearErrors', () => {
-    it('should remove error', () => {
-      const { result } = renderHook(() => useForm<{ input: string }>());
-      act(() => {
-        result.current.register('input');
-        result.current.setError('input', {
-          type: 'test',
-          message: 'message',
-        });
-      });
-
-      act(() => result.current.clearErrors('input'));
-
-      expect(result.current.formState.errors).toEqual({});
-    });
-
-    it('should remove nested error', () => {
-      const { result } = renderHook(() =>
-        useForm<{ input: { nested: string } }>(),
-      );
-      act(() =>
-        result.current.setError('input.nested', {
-          type: 'test',
-        }),
-      );
-      expect(result.current.formState.errors.input?.nested).toBeDefined();
-      act(() => result.current.clearErrors('input.nested'));
-      expect(result.current.formState.errors.input?.nested).toBeUndefined();
-    });
-
-    it('should remove deep nested error and set it to undefined', async () => {
-      let currentErrors = {};
-
-      const Component = () => {
-        const {
-          register,
-          formState: { errors },
-          trigger,
-          clearErrors,
-        } = useForm<{
-          test: { data: string };
-        }>();
-
-        currentErrors = errors;
-        return (
-          <div>
-            <input type="text" {...register('test.data', { required: true })} />
-            <button type={'button'} onClick={() => trigger()}>
-              submit
-            </button>
-            <button type={'button'} onClick={() => clearErrors(['test.data'])}>
-              clear
-            </button>
-          </div>
-        );
-      };
-
-      await actComponent(async () => {
-        render(<Component />);
-      });
-
-      await actComponent(async () => {
-        fireEvent.click(screen.getByRole('button', { name: 'submit' }));
-      });
-
-      expect(currentErrors).toMatchSnapshot();
-
-      await actComponent(async () => {
-        fireEvent.click(screen.getByRole('button', { name: 'clear' }));
-      });
-
-      expect(currentErrors).toEqual({});
-    });
-
-    it('should remove specified errors', () => {
-      const { result } = renderHook(() =>
-        useForm<{
-          input: string;
-          input1: string;
-          input2: string;
-          nest: { data: string; data1: string };
-        }>(),
-      );
-
-      const error = {
-        type: 'test',
-        message: 'message',
-      };
-
-      act(() => {
-        result.current.register('input');
-        result.current.register('input1');
-        result.current.register('input2');
-        result.current.setError('input', error);
-        result.current.setError('input1', error);
-        result.current.setError('input2', error);
-
-        result.current.register('nest.data');
-        result.current.register('nest.data1');
-        result.current.setError('nest.data', error);
-        result.current.setError('nest.data1', error);
-      });
-
-      const errors = {
-        input: {
-          ...error,
-          ref: {
-            name: 'input',
-          },
-        },
-        input1: {
-          ...error,
-          ref: {
-            name: 'input1',
-          },
-        },
-        input2: {
-          ...error,
-          ref: {
-            name: 'input2',
-          },
-        },
-        nest: {
-          data: {
-            ...error,
-            ref: {
-              name: 'nest.data',
-            },
-          },
-          data1: {
-            ...error,
-            ref: {
-              name: 'nest.data1',
-            },
-          },
-        },
-      };
-      expect(result.current.formState.errors).toEqual(errors);
-
-      act(() => result.current.clearErrors(['input', 'input1', 'nest.data']));
-      expect(result.current.formState.errors).toEqual({
-        input2: errors.input2,
-        nest: {
-          data1: errors.nest.data1,
-        },
-      });
-    });
-
-    it('should remove all error', () => {
-      const { result } = renderHook(() =>
-        useForm<{ input: string; input1: string; input2: string }>(),
-      );
-
-      const error = {
-        type: 'test',
-        message: 'message',
-      };
-      act(() => result.current.setError('input', error));
-      act(() => result.current.setError('input1', error));
-      act(() => result.current.setError('input2', error));
-      expect(result.current.formState.errors).toEqual({
-        input: {
-          ...error,
-          ref: undefined,
-          types: undefined,
-        },
-        input1: {
-          ...error,
-          ref: undefined,
-          types: undefined,
-        },
-        input2: {
-          ...error,
-          ref: undefined,
-          types: undefined,
-        },
-      });
-
-      act(() => result.current.clearErrors());
-      expect(result.current.formState.errors).toEqual({});
-    });
-
-    it('should prevent the submission if there is a custom error', async () => {
-      const submit = jest.fn();
-      const { result } = renderHook(() =>
-        useForm<{ data: string; whatever: string }>(),
-      );
-
-      result.current.register('data');
-
-      act(() => {
-        result.current.setError('whatever', { type: 'server' });
-      });
-
-      await act(async () => await result.current.handleSubmit(submit)());
-      expect(submit).not.toBeCalled();
-
-      act(() => {
-        result.current.clearErrors('whatever');
-      });
-
-      await act(async () => await result.current.handleSubmit(submit)());
-      expect(submit).toBeCalled();
-    });
-  });
-
-  describe('formState', () => {
-    it('should return isValid correctly with resolver', async () => {
-      let isValidValue = false;
-
-      const Component = () => {
-        const {
-          register,
-          formState: { isValid },
-        } = useForm<{ test: string }>({
-          mode: 'onChange',
-          resolver: async (data) => {
-            return {
-              values: data.test ? data : {},
-              errors: data.test
-                ? {}
-                : {
-                    test: {
-                      message: 'issue',
-                      type: 'test',
-                    },
-                  },
-            };
-          },
-        });
-
-        isValidValue = isValid;
-        return <input {...register('test')} />;
-      };
-
-      await actComponent(async () => {
-        render(<Component />);
-      });
-
-      expect(isValidValue).toBeFalsy();
-
-      await actComponent(async () => {
-        fireEvent.input(screen.getByRole('textbox'), {
-          target: {
-            value: 'test',
-          },
-        });
-      });
-
-      await actComponent(async () => {
-        expect(isValidValue).toBeTruthy();
-      });
-    });
-
-    it('should return true for onBlur mode by default', () => {
-      const { result } = renderHook(() =>
-        useForm<{ input: string }>({
-          mode: VALIDATION_MODE.onBlur,
-        }),
-      );
-
-      expect(result.current.formState.isValid).toBeTruthy();
-    });
-
-    it('should return true for onChange mode by default', () => {
-      const { result } = renderHook(() =>
-        useForm<{ input: string }>({
-          mode: VALIDATION_MODE.onChange,
-        }),
-      );
-
-      expect(result.current.formState.isValid).toBeTruthy();
-    });
-
-    it('should return true when no validation is registered', () => {
-      const { result } = renderHook(() =>
-        useForm<{ test: string }>({
-          mode: VALIDATION_MODE.onBlur,
-        }),
-      );
-
-      result.current.register('test');
-
-      expect(result.current.formState.isValid).toBeTruthy();
-    });
-
-    it('should return false when default value is not valid value', async () => {
-      const { result } = renderHook(() => {
-        const methods = useForm<{ input: string; issue: string }>({
-          mode: VALIDATION_MODE.onChange,
-        });
-
-        methods.formState.isValid;
-
-        return methods;
-      });
-
-      await act(async () => {
-        result.current.register('issue', { required: true });
-        result.current.setValue('issue', '', { shouldValidate: true });
-      });
-
-      expect(result.current.formState.isValid).toBeFalsy();
-    });
-
-    it('should return true when default value meet the validation criteria', async () => {
-      const { result } = renderHook(() =>
-        useForm<{ input: string; issue: string }>({
-          mode: VALIDATION_MODE.onChange,
-        }),
-      );
-
-      result.current.formState.isValid;
-
-      await act(async () => {
-        result.current.register('issue', { required: true });
-      });
-
-      expect(result.current.formState.isValid).toBeTruthy();
-    });
-
-    it('should be a proxy object that returns undefined for unknown properties', () => {
-      const { result } = renderHook(() => useForm());
-
-      // @ts-expect-error
-      expect(result.current.formState.nonExistentProperty).toBeUndefined();
-    });
-
-    it('should be a proxy object that properly implements the has trap', () => {
-      const { result } = renderHook(() => useForm());
-
-      expect('nonExistentProperty' in result.current.formState).toBeFalsy();
-    });
-
-    it('should be a proxy object that hasOwnProperty works on', () => {
-      const { result } = renderHook(() => useForm());
-
-      expect(result.current.formState).toHaveProperty('hasOwnProperty');
     });
   });
